Added Debian Info
I've added info to the README regarding how to install requirements and run the script on a Debian 10 system. Because Python 2 and Python 3 are installed alongside each other, all dependencies needs to be specifically installed for Python 3 and Python 3 called to run the script.

I thought this info might save 30 minutes in someone else's life. Feel free to reject the pull request if you think it's not worth being this specific for one operating system.